+++ src/snowflake/snowpark/dataframe_reader.py
@@ -434,7 +434,6 @@ class DataFrameReader:
                             path,
                             self._file_type,
                             self._cur_options,
-                            self._session.get_fully_qualified_current_schema(),
                             schema,
                             schema_to_cast=schema_to_cast,
                             transformations=transformations,
@@ -453,7 +452,6 @@ class DataFrameReader:
                     path,
                     self._file_type,
                     self._cur_options,
-                    self._session.get_fully_qualified_current_schema(),
                     schema,
                     schema_to_cast=schema_to_cast,
                     transformations=transformations,
@@ -576,10 +574,8 @@ class DataFrameReader:
     ) -> Tuple[List, List, List, Exception]:
         format_type_options, _ = get_copy_into_table_options(self._cur_options)
 
-        temp_file_format_name = (
-            self._session.get_fully_qualified_current_schema()
-            + "."
-            + random_name_for_temp_object(TempObjectType.FILE_FORMAT)
+        temp_file_format_name = self._session.get_fully_qualified_name_if_possible(
+            random_name_for_temp_object(TempObjectType.FILE_FORMAT)
         )
         drop_tmp_file_format_if_exists_query: Optional[str] = None
         use_temp_file_format = "FORMAT_NAME" not in self._cur_options
@@ -685,7 +681,6 @@ class DataFrameReader:
                             path,
                             format,
                             self._cur_options,
-                            self._session.get_fully_qualified_current_schema(),
                             schema,
                             schema_to_cast=schema_to_cast,
                             transformations=read_file_transformations,
@@ -704,7 +699,6 @@ class DataFrameReader:
                     path,
                     format,
                     self._cur_options,
-                    self._session.get_fully_qualified_current_schema(),
                     schema,
                     schema_to_cast=schema_to_cast,
                     transformations=read_file_transformations,